Refactoring for eventhub data writer (#1787)
* Refactoring for eventhub data writer

* Add more comments

* Rename accumulator and batch
